DAOS-654 pool: Modify pool connect to use security credential
This patch modifies the pool connect RPC to pass the opaque security credential
instead of uid and gid of the calling process. It also modifies the daos client
library call for pool connect to request the opaque credential from the agent
and packs it into the connect RPC. on the server end the pool connect handler
has been modified to pass the credential and the pool attributes which contain
the pool ownership information to a new can_pool_connect api call in the
security module.

Change-Id: If512ac64d4787747c88df82ff4ff63bafae18edf
Signed-off-by: David Quigley <david.quigley@intel.com>